[added] Particle emitter test.
32x32 image
300 particles: 41fps droid, 26fps g1
150 particles: 49fps droid, 41fps g1
80 particles: 52fps droid, 55-60fps g1

64x64 image, rotation, color change
150 particles: 31fps droid, 25fps g1
80 particles: 40fps droid, 37fps g1

128x128 image, scaled to 80x80, rotation, color change
150 particles: 13fps droid, 11fps g1
80 particles: 46fps droid, 55-60fps g1

128x128, scaled to 256x256, rotation, color change
10 particles: droid 38fps, g1 37fps
10 particles: droid 46fps, g1 55-60fps

+++ gdx/src/com/badlogic/gdx/graphics/particles/ParticleEmitter.java
@@ -833,12 +833,12 @@ public class ParticleEmitter {
 			return lowMin + (lowMax - lowMin) * MathUtils.random();
 		}
 
-		public void setLow (int value) {
+		public void setLow (float value) {
 			lowMin = value;
 			lowMax = value;
 		}
 
-		public void setLow (int min, int max) {
+		public void setLow (float min, float max) {
 			lowMin = min;
 			lowMax = max;
 		}
@@ -890,12 +890,12 @@ public class ParticleEmitter {
 			return highMin + (highMax - highMin) * MathUtils.random();
 		}
 
-		public void setHigh (int value) {
+		public void setHigh (float value) {
 			highMin = value;
 			highMax = value;
 		}
 
-		public void setHigh (int min, int max) {
+		public void setHigh (float min, float max) {
 			highMin = min;
 			highMax = max;
 		}
